chore: fix warning (#173)
<!-- Thank you for your contribution to samurai! -->

<!-- Please check the following before submitting your PR -->
- [ ] I have installed [pre-commit](https://pre-commit.com/) locally and
use it to validate my commits.
- [ ] The PR title follows the [conventional
commits](https://www.conventionalcommits.org/en/v1.0.0/) convention.
Available tags: 'build', 'chore', 'ci', 'docs', 'feat', 'fix', 'perf',
'refactor', 'revert', 'style', 'test'
- [ ] This new PR is documented.
- [ ] This new PR is tested.

## Description
<!-- A clear and concise description of what you have done in this PR.
-->

## Related issue
<!-- List the issues solved by this PR, if any. -->

## How has this been tested?
<!-- Give the list of files used to test this new implementation. -->

## Code of Conduct
By submitting this PR, you agree to follow our [Code of
Conduct](https://github.com/hpc-maths/samurai/blob/master/docs/CODE_OF_CONDUCT.md)
- [ ] I agree to follow this project's Code of Conduct

+++ include/samurai/stencil.hpp
@@ -298,7 +298,7 @@ namespace samurai
 
         Stencil<stencil_size, dim> s;
         s.fill(0);
-        static_for<0, stencil_size>::apply( // for (int i=0; i<stencil_size; i++)
+        static_for<0, static_cast<int>(stencil_size)>::apply( // for (int i=0; i<stencil_size; i++)
             [&](auto integral_constant_i)
             {
                 static constexpr int i = decltype(integral_constant_i)::value;
@@ -313,7 +313,7 @@ namespace samurai
     {
         Stencil<stencil_size, dim> s;
         s.fill(0);
-        static_for<0, stencil_size>::apply( // for (int i=0; i<stencil_size; i++)
+        static_for<0, static_cast<int>(stencil_size)>::apply( // for (int i=0; i<stencil_size; i++)
             [&](auto integral_constant_i)
             {
                 static constexpr int i = decltype(integral_constant_i)::value;
